Another interim checkin *however* loosely things now work.
There is a rather large issue that the strategy instances are currently re-used so will
be in an in-consistent state of the node server is accessing more than one authorized route
at a time.

Need to still fix the 'race', secure the session cookie, and generally fix the bazillion issues!
(Oh and write more specs.)