allow forcing the display of the OAuth approval screen
as per https://developers.google.com/accounts/docs/OAuth2WebServer#offline the only way to regenerate a refresh_token (for offline use), is to force the user to re-approve the app. This commit adds the optional "forceApproval" option. If present and is true, it will trigger the display of the approval prompt to the user.